requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Celery и Redis
celery==5.3.4
redis==5.0.1

# Instagram
instagrapi==2.0.0
# Примечание: instagrapi требует pydantic==1.10.9, но мы используем pydantic 2.x
# Установим pydantic 1.10.9 для совместимости с instagrapi
pydantic==1.10.9
Pillow>=8.1.1  # Требуется для instagrapi

# Переводы
openai==1.3.0  # Для DeepSeek API (OpenAI-совместимый)

# Безопасность
cryptography==41.0.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Валидация
# pydantic==2.5.0  # Закомментировано из-за конфликта с instagrapi
# pydantic-settings==2.1.0  # Требует pydantic 2.x, используем BaseSettings из pydantic напрямую

# Утилиты
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Мониторинг (опционально)
sentry-sdk[fastapi]==1.38.0

